Adding important homework files to Github
Using library computer, so I gotta finish up by adding the final files. In here should be the plaintext of the text to cipher, the key.txt, and the ciphered code (onepunch.man). Homework.py is the final prompt that I need to fill out for the teacher. Otherwise we should be set! Almost done :) 

*Note to self: the teacher will be testing the code by ensuring that the original plaintext and the deciphered text will be the same using the key provided. She will also try to cipher the plaintext herself and decipher it using a random key (no key_name input).